% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kat_fact.R
\name{kat_fact}
\alias{kat_fact}
\title{get a cat fact}
\usage{
kat_fact(max_length = NULL, ...)
}
\arguments{
\item{max_length}{(integer) limits the length of the facts returned}

\item{...}{curl options passed on to \code{httr::GET}}
}
\value{
a cat fact in a list
}
\description{
get a cat fact
}
\examples{
\dontrun{
kat_fact()
kat_fact(40)
kat_fact(80)
# kat_fact('asdf') # invalid max_length returns nothing
}
}
